requirements.txt
aiodns==4.0.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
backoff==2.2.1
backports.zstd==1.3.0
bcrypt==5.0.0
boto3==1.42.36
botocore==1.42.36
brotli==1.2.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==45.0.7
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.128.0
fastapi-cli==0.0.20
fastapi-cloud-cli==0.11.0
fastar==0.8.0
filelock==3.20.3
frozenlist==1.8.0
fsspec==2026.2.0
grpcio==1.78.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.4.1
hyperframe==6.1.0
idna==3.11
inquirerpy==0.3.4
invoke==2.2.1
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.1.0
joblib==1.5.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
networkx==3.6.1
numpy==2.4.2
orjson==3.11.5
packaging==26.0
paramiko==4.0.0
pfzy==0.3.4
pillow==12.1.0
portalocker==3.2.0
prettytable==3.17.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.5
py-cpuinfo==9.0.0
pycares==5.0.1
pycparser==3.0
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyNaCl==1.6.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
PyYAML==6.0.3
qdrant-client==1.16.2
regex==2026.2.19
requests==2.32.5
rich==14.3.1
rich-toolkit==0.17.1
rignore==0.7.6
ruff==0.14.13
runpod==1.8.1
s3transfer==0.16.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
sentence-transformers==5.2.3
sentry-sdk==2.51.0
setuptools==82.0.0
shellingham==1.5.4
six==1.17.0
starlette==0.50.0
sympy==1.14.0
threadpoolctl==3.6.0
timm==1.0.24
tokenizers==0.22.2
tomli==2.4.0
tomlkit==0.14.0
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.1
tqdm-loggable==0.2
transformers==5.2.0
typer==0.24.1
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
ujson==5.11.0
urllib3==2.6.3
uvicorn==0.40.0
uvloop==0.22.1
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.5.0
websockets==16.0
yarl==1.22.0
